134 - roles

+++ src/main/java/com/febfes/fftmback/controller/ProjectController.java
@@ -1,12 +1,16 @@
 package com.febfes.fftmback.controller;
 
 import com.febfes.fftmback.annotation.*;
+import com.febfes.fftmback.config.auth.RoleCheckerComponent;
+import com.febfes.fftmback.domain.common.RoleName;
 import com.febfes.fftmback.domain.dao.ProjectEntity;
 import com.febfes.fftmback.domain.dao.TaskTypeEntity;
 import com.febfes.fftmback.domain.dao.UserEntity;
+import com.febfes.fftmback.dto.OneProjectDto;
 import com.febfes.fftmback.dto.PatchDto;
 import com.febfes.fftmback.dto.ProjectDto;
 import com.febfes.fftmback.dto.UserDto;
+import com.febfes.fftmback.exception.ProjectOwnerException;
 import com.febfes.fftmback.mapper.ProjectMapper;
 import com.febfes.fftmback.mapper.UserMapper;
 import com.febfes.fftmback.service.ProjectService;
@@ -34,6 +38,7 @@ public class ProjectController {
 
     private final @NonNull ProjectService projectService;
     private final @NonNull TaskTypeService taskTypeService;
+    private final @NonNull RoleCheckerComponent roleCheckerComponent;
 
     @Operation(summary = "Get all projects for authenticated user")
     @ApiGet
@@ -58,9 +63,9 @@ public class ProjectController {
     @Operation(summary = "Get project by its id")
     @ApiGetOne(path = "{id}")
     @SuppressWarnings("MVCPathVariableInspection") // fake warn "Cannot resolve path variable 'id' in @RequestMapping"
-    public ProjectDto getProject(@PathVariable Long id) {
+    public OneProjectDto getProject(@PathVariable Long id) {
         UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return ProjectMapper.INSTANCE.projectToProjectDto(projectService.getProjectForUser(id, user.getId()));
+        return projectService.getProjectForUser(id, user.getId());
     }
 
     @Operation(summary = "Edit project by its id")
@@ -69,12 +74,14 @@ public class ProjectController {
             @PathVariable Long id,
             @RequestBody ProjectDto projectDto
     ) {
+        roleCheckerComponent.checkIfHasRole(id, RoleName.MEMBER_PLUS);
         projectService.editProject(id, ProjectMapper.INSTANCE.projectDtoToProject(projectDto));
     }
 
     @Operation(summary = "Delete project by its id")
     @ApiDelete(path = "{id}")
     public void deleteProject(@PathVariable Long id) {
+        roleCheckerComponent.checkIfHasRole(id, RoleName.OWNER);
         projectService.deleteProject(id);
     }
 
@@ -103,8 +110,7 @@ public class ProjectController {
     @ApiResponse(responseCode = "404", description = "Project not found", content = @Content)
     @ApiResponse(responseCode = "409", description = "Only owner can add a new member", content = @Content)
     public List<UserDto> addNewMembers(@PathVariable Long id, @RequestBody List<Long> memberIds) {
-        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        memberIds.forEach(memberId -> projectService.projectOwnerCheck(id, user.getId()));
+        roleCheckerComponent.checkIfHasRole(id, RoleName.MEMBER_PLUS);
         List<UserEntity> addedMembers = projectService.addNewMembers(id, memberIds);
         return addedMembers.stream()
                 .map(UserMapper.INSTANCE::userToUserDto)
@@ -116,8 +122,10 @@ public class ProjectController {
     @ApiResponse(responseCode = "404", description = "Project not found", content = @Content)
     @ApiResponse(responseCode = "409", description = "Only owner can remove a member", content = @Content)
     public UserDto removeMember(@PathVariable Long id, @PathVariable Long memberId) {
-        UserEntity user = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        projectService.projectOwnerCheck(id, user.getId());
+        roleCheckerComponent.checkIfHasRole(id, RoleName.MEMBER_PLUS);
+        if (roleCheckerComponent.userHasRole(id, memberId, RoleName.OWNER)) {
+            throw new ProjectOwnerException();
+        }
         UserEntity removedMember = projectService.removeMember(id, memberId);
         return UserMapper.INSTANCE.userToUserDto(removedMember);
     }
